+++ src/components/tag-section/index.ts
@@ -1,0 +1,3 @@
+import TagSection from './tag-section';
+
+export default TagSection;
