+++ src/components/tag-section/styles.ts
@@ -1,0 +1,12 @@
+export const tagWrapperStyles = {
+    display: ['none', null, 'flex'],
+    gap: 'inc20',
+    flexWrap: 'wrap' as 'wrap',
+};
+
+export const tagStyles = {
+    display: ['none', null, 'block'],
+    px: [null, null, null, 'inc30'],
+    py: [null, null, null, 'inc20'],
+    fontSize: [null, null, null, 'inc20'],
+};
