+++ src/components/tag-section/types.ts
@@ -1,0 +1,3 @@
+export interface TagSectionProps {
+    tags: string[];
+}
